Factor out an uploadMeteorTestApp command.
I'd wanted to do this previously, but couldn't figure out where these
were defined. Now we know.
