+++ tests/tests/account-settings.js
@@ -88,24 +88,8 @@ module.exports["Test profile changes passing to testapp"] = function (browser) {
     })
 
 
-    // Navigate to apps page, upload meteor-testapp.spk, create new instance
-    .click('a[href="/apps"]')
-    .waitForElementVisible('.upload-button', short_wait)
-    .execute(function () {
-      document.querySelector("input[type=file]")
-              .setAttribute("style", ""); 
-      }, [])
-    .perform(function (client, done) {
-      client.setValue("input[type=file]", testappPath, () => {
-        done();
-      })
-    })
-    .execute(function () {
-      const testappFile = document.querySelector("input[type=file]").files[0];
-      uploadApp(testappFile);
-    }, [])
-    .waitForElementVisible('#confirmInstall', long_wait)
-    .click('#confirmInstall')
+    // upload meteor-testapp.spk, create new instance
+    .uploadMeteorTestApp()
     .waitForElementVisible('button.action', medium_wait)
     .click('button.action')
