+++ tests/tests/appHooks.js
@@ -15,61 +15,41 @@
 // limitations under the License.
 
 const {short_wait, medium_wait, long_wait} = require('../utils');
-const testappPath = require('path').resolve(__dirname + "/../assets/meteor-testapp.spk");
-
-// TODO: move this out into some shared module, and use it account-settings.js too.
-const installMeteorTestApp = (chain) => {
-  return chain
-    .url('/apps')
-    .waitForElementVisible('.upload-button', short_wait)
-    .perform(function (client, done) {
-      client.setValue("input[type=file]", testappPath, () => {
-        done();
-      })
-    })
-    .execute(function () {
-      const testappFile = document.querySelector("input[type=file]").files[0];
-      uploadApp(testappFile);
-    }, [])
-    .waitForElementVisible('#confirmInstall', long_wait)
-    .click('#confirmInstall')
-}
 
 module.exports['Test saving and restoring capabilities via AppHooks'] = function(browser) {
-  installMeteorTestApp(
-    browser
-      .loginDevAccount()
-  )
-  .waitForElementVisible('button.action', medium_wait)
-  .click('button.action')
-  // click the "Got it!" button in the "this is your first grain" popup.
-  // This shadows the debug log so we have to clear it in order to open the
-  // log.
-  .waitForElementVisible(".introjs-skipbutton", short_wait)
-  .click(".introjs-skipbutton")
-  .waitForElementVisible("#openDebugLog", short_wait)
-  // We also want to wait a moment before clicking the debug log button, so the
-  // first grain popup has actually gotten out of our way:
-  .pause(short_wait)
-  .click("#openDebugLog")
-  .waitForElementVisible('.grain-frame', medium_wait)
-  .grainFrame()
-  .execute(function() {
-    Meteor.call('schedule', 'someObjectId')
-  })
-  .frameParent()
-  .pause(short_wait)
-  .execute(function() {
-    Meteor.call("runDueJobsAt", Date.now() + 1000 * 60 * 60);
-  })
-  .pause(short_wait)
-  .windowHandles(windows => browser.switchWindow(windows.value[1]))
-  .assert.containsText(".grainlog-contents > pre", "Running callback: someObjectId")
+  browser
+    .loginDevAccount()
+    .uploadMeteorTestApp()
+    .waitForElementVisible('button.action', medium_wait)
+    .click('button.action')
+    // click the "Got it!" button in the "this is your first grain" popup.
+    // This shadows the debug log so we have to clear it in order to open the
+    // log.
+    .waitForElementVisible(".introjs-skipbutton", short_wait)
+    .click(".introjs-skipbutton")
+    .waitForElementVisible("#openDebugLog", short_wait)
+    // We also want to wait a moment before clicking the debug log button, so the
+    // first grain popup has actually gotten out of our way:
+    .pause(short_wait)
+    .click("#openDebugLog")
+    .waitForElementVisible('.grain-frame', medium_wait)
+    .grainFrame()
+    .execute(function() {
+      Meteor.call('schedule', 'someObjectId')
+    })
+    .frameParent()
+    .pause(short_wait)
+    .execute(function() {
+      Meteor.call("runDueJobsAt", Date.now() + 1000 * 60 * 60);
+    })
+    .pause(short_wait)
+    .windowHandles(windows => browser.switchWindow(windows.value[1]))
+    .assert.containsText(".grainlog-contents > pre", "Running callback: someObjectId")
 
-  // Close the grain log and switch back to the main window, to avoid
-  // confusing future tests:
-  browser.windowHandles(windows => {
-    browser.closeWindow()
-    browser.switchWindow(windows.value[0])
-  })
+    // Close the grain log and switch back to the main window, to avoid
+    // confusing future tests:
+    browser.windowHandles(windows => {
+      browser.closeWindow()
+      browser.switchWindow(windows.value[0])
+    })
 }
